Improve redo setup, bad target files.

+++ all.do
@@ -4,9 +4,9 @@
 # arguments.
 
 # Remove target (html) files for which no sources files can be found.
-IFS=$(echo -en "\n\b")
-for file in `find . -name "*.html"`; do
-  if ! test -f "${file%.html}.rst" &&
+for file in *.html; do
+  if   test -f "$file" &&
+     ! test -f "${file%.html}.rst" &&
      ! test -f "${file%.html}.md"; then
     rm "$file"
   fi
@@ -14,9 +14,19 @@ done
 
 # Determine target files from the sources files present, declare dependencies
 # of the all.do script on them / build them if necessary.
-for file in `find . -name "*.rst"`; do
+for file in *.rst; do
   redo-ifchange "${file%.rst}.html"
 done
-for file in `find . -name "*.md"`; do
+for file in *.md; do
   redo-ifchange "${file%.md}.html"
+done
+
+# Run redo in subdirectories, and copy current all.do there if none exists.
+for d in *; do
+  if test -d "$d"; then
+    if ! test -f "$d"/all.do; then
+      cp $0 "$d"/all.do
+    fi
+    redo "$d"/all
+  fi
 done
